# contracts/openapi.yaml
openapi: 3.1.0
...
info:
  title: JourneyLab API
  version: 0.1.0
  summary: Trip digital twin — compare feasible futures before and during travel.
  description: |
    Contract-first. Nothing in this document is implemented yet; every operation
    is `PROPOSED` until its sub-step lands.

    **Deterministic engines own feasibility; the model owns language** (`ADR-002`).
    No endpoint returns a plan the constraint engine has not validated.
  license:
    name: UNLICENSED
    identifier: LicenseRef-proprietary
servers:
  - url: https://api.journeylab.app/v1
    description: Production
  - url: https://localhost:5709/v1
    description: Local development

# The major version is in the path, not a header. `API_CONTRACTS.md` §1: a
# breaking change means a new major version, and a version a caller can forget to
# send is a version they will forget to send.
...
paths:
...
  /trips:
    post:
      operationId: createTrip
      tags: [trips]
      summary: Create a trip and its initial planning context.
      description: |
        Authenticated user **or a valid guest session**; creates ownership.

        A guest can plan. That is a product decision, not an oversight: requiring
        an account before a traveller has seen anything useful is the most
        reliable way to lose them. The guest session lasts 7 days (`ADR-013`).
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTripRequest' }
            example:
              origin: { name: 'London', time_zone: 'Europe/London' }
              destination_region: 'gr-cyclades'
              date_range: { start: '2026-09-12', end: '2026-09-19' }
              party: { adults: 2, children: [8], accessibility_needs: [] }
              currency: 'EUR'
              locale: 'en-GB'
      responses:
        '201':
          description: Trip created.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Trip' }
        '400': { $ref: '#/components/responses/Problem' }
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          description: |
            Outside coverage. **A product state, not a failure** — the response
            names the supported regions or date bounds so the traveller can act.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
              example:
                type: 'https://journeylab.app/problems/coverage.unsupported_region'
                title: 'Region not in the destination pack'
                status: 422
                code: 'coverage.unsupported_region'
                correlation_id: 'corr_9f3c2a1b4d5e6f70'
                retryable: false
                detail: 'The Cyclades are not in the Phase 1 destination pack.'
                remediation:
                  kind: 'choose_supported_region'
        '429': { $ref: '#/components/responses/RateLimited' }

  # ==========================================================================
  # API-002 — read a trip
  # ==========================================================================
...
  /coverage:
    get:
      operationId: getCoverage
      tags: [platform]
      summary: Supported regions, date bounds, freshness and documented limits.
      # The ONLY unauthenticated operation in this contract. Declared explicitly
      # rather than inherited, so it is visible in a diff.
      security: []
      description: |
        **Public and unauthenticated** — the one operation in this contract that
        is. A traveller must be able to learn whether their destination is
        supported before creating an account, and putting that behind a login
        means asking someone to register to be told "no".

        **It must never expose provider identities or quota details**
        (`REQ-EVID-006`). Which supplier backs a region is commercially
        confidential; how close a provider is to its quota tells an attacker
        exactly when the product degrades. Coverage answers *what is supported*,
        never *how it is supplied*.

        `provider_health` is therefore an aggregate label, not a per-provider
        breakdown.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: Coverage.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Coverage' }
              example:
                regions:
                  - region_id: 'gr-cyclades'
                    display_name: 'Cyclades, Greece'
                    date_bounds: { start: '2026-04-01', end: '2027-03-31' }
                    freshness: 'current'
                    limitations: ['Ferry timetables are seasonal and change monthly']
                provider_health: 'healthy'
                # When the answer was read, not when it was requested — a cached response
                # carries the time of the read that filled the cache (STEP-007.05).
                observed_at: '2026-09-14T08:00:00Z'
        '429': { $ref: '#/components/responses/RateLimited' }

  # ==========================================================================
  # API-019 — can this request be planned at all
  # ==========================================================================
...
  /waitlist:
    post:
      operationId: joinWaitlist
      tags: [platform]
      summary: Ask to be told when an unsupported destination becomes plannable.
      # Public for the same reason the two coverage operations are: this is the
      # thing a visitor does immediately after being told no, and asking them to
      # register first is asking them to register in order to be kept informed
      # about a product they cannot yet use.
      security: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      description: |
        **Public and unauthenticated**, and the first operation in this contract
        that writes a stranger's personal data.

        **Consent is an action, not a default.** `consent.granted` has no default
        in the schema, no default in the request model and no default in the
        database. A request that omits it fails validation; a request that sends
        `false` is refused with the field named and nothing is stored. There is no
        path on which an entry is written because consent was assumed.

        **The address is never echoed.** A validation failure names the offending
        *fields* and never their contents — `BUG-035` was precisely this, an
        unauthenticated endpoint reflecting the value that failed back to whoever
        sent it, which on this operation would be somebody's email address.

        **`withdrawal_token` is returned exactly once and never again.** Only its
        hash is stored, so it cannot be re-read, re-sent or recovered — it is the
        entire authorisation for `withdrawWaitlistConsent`. Email delivery is not
        part of this sub-step, so there is nothing to send a link through; the
        token is what makes withdrawal exercisable today by somebody who has no
        account and no session.

        **The response does not reveal whether the address was already listed.**
        A repeated request returns the same shape with a fresh token rather than a
        conflict, so this operation cannot be used to test whether a given address
        is on the list.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WaitlistJoinRequest' }
            example:
              email: 'traveller@example.com'
              region_query: 'Faroe Islands'
              consent:
                purpose: 'waitlist_notification'
                granted: true
      responses:
        '201':
          description: |
            Recorded. `withdrawal_token` appears here and nowhere else, ever.
          headers:
            X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WaitlistJoined' }
              example:
                purpose: 'waitlist_notification'
                basis: 'consent'
                granted_at: '2026-09-10T09:24:11Z'
                withdrawal_token: 'K8s1_o0Wq3nR2vF7pZ4bLyX6dTgH9cJmA0eQnU5tYwI'
        '400':
          description: |
            The request does not satisfy the contract, or consent was not granted.
            The offending fields are named; their values are not.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
              example:
                type: 'https://journeylab.app/problems/validation.invalid_request'
                title: 'Request does not satisfy its schema'
                status: 400
                code: 'validation.invalid_request'
                correlation_id: 'corr_9f3c2a1b4d5e6f70'
                retryable: false
                detail: >-
                  A waitlist entry requires consent to be contacted for that
                  purpose. Nothing has been stored.
                remediation:
                  kind: 'correct_fields'
                  fields: ['consent.granted']
        '503': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
components:
...
  schemas:
...
    CreateTripRequest:
      type: object
      required: [destination_region, date_range, party, currency, locale]
      additionalProperties: false
      properties:
        origin: { $ref: '#/components/schemas/Place' }
        destination_region: { type: string }
        date_range: { $ref: '#/components/schemas/LocalDateRange' }
        party: { $ref: '#/components/schemas/Party' }
        currency: { type: string, pattern: '^[A-Z]{3}$' }
        locale: { type: string }
    Trip:
      type: object
      required: [trip_id, destination_region, date_range, party, currency, locale, version]
      properties:
        trip_id: { type: string }
        origin: { $ref: '#/components/schemas/Place' }
        destination_region: { type: string }
        date_range: { $ref: '#/components/schemas/LocalDateRange' }
        party: { $ref: '#/components/schemas/Party' }
        currency: { type: string, pattern: '^[A-Z]{3}$' }
        locale: { type: string }
        version: { type: integer, minimum: 1 }
        canonical_scenario_id:
          oneOf:
            - { type: string }
            - { type: 'null' }
          description: |
            Exactly zero or one (`DATA-004`). A trip with two canonical plans is
            a trip nobody has decided.
        permissions:
          type: array
          description: |
            What the CALLER may do. Presentation only — the server enforces
            regardless, and a client that hides a button has hidden a button, not
            prevented an action.
          items: { type: string }
...
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        **Required on every state-changing command.** Optional idempotency is
        idempotency nobody uses until after the duplicate-charge incident.
        Generate one per user intent and reuse it for every retry of that intent;
        a replay returns the original result, and reuse with a different body is
        a client defect and returns 409.
      schema:
        type: string
        pattern: '^[A-Za-z0-9._:\-]{8,255}$'
...
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      description: Echoed back. Minted when absent or unusable.
      schema:
        type: string
        pattern: '^[A-Za-z0-9._:\-]{8,128}$'
...
  headers:
    ETag:
      description: |
        Strong entity tag. Strong rather than weak (`W/`) because `If-Match`
        requires strong comparison — a weak tag makes the precondition succeed for
        representations that are equivalent but not identical, which is exactly
        the case a lost-update check must reject.
      required: true
      schema: { type: string }
    CorrelationId:
      description: Echoed from the request, or minted when absent or unusable.
      required: true
      schema: { type: string }
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        OIDC/OAuth 2.1, short-lived. **Tenant context is derived from the token,
        never from a client-supplied header or body field** (`REQ-SEC-001`).

  # --- reusable responses ---------------------------------------------------------
...
security:
  - bearerAuth: []